Move Quaternion and Vector operations to math folder
