+++ src/restruct_poc/math/QuaternionOperations.hpp
