+++ src/restruct_poc/math/VectorOperations.hpp
