+++ src/restruct_poc/solver/CalculateConstraintResidualGradient.hpp
@@ -4,7 +4,7 @@
 
 #include "Constraints.hpp"
 
-#include "src/restruct_poc/QuaternionOperations.hpp"
+#include "src/restruct_poc/math/QuaternionOperations.hpp"
 #include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
