+++ src/restruct_poc/solver/CalculateDisplacement.hpp
@@ -2,9 +2,8 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "src/restruct_poc/QuaternionOperations.hpp"
+#include "src/restruct_poc/math/QuaternionOperations.hpp"
 #include "src/restruct_poc/types.hpp"
-#include "src/restruct_poc/QuaternionOperations.hpp"
 
 namespace openturbine {
