+++ src/restruct_poc/system/CalculateInertiaStiffnessMatrix.hpp
@@ -5,7 +5,7 @@
 #include <KokkosBlas1_set.hpp>
 #include <Kokkos_Core.hpp>
 
-#include "src/restruct_poc/VectorOperations.hpp"
+#include "src/restruct_poc/math/VectorOperations.hpp"
 #include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
