+++ src/restruct_poc/system/CalculateInertialForces.hpp
@@ -4,7 +4,7 @@
 #include <KokkosBlas.hpp>
 #include <Kokkos_Core.hpp>
 
-#include "src/restruct_poc/VectorOperations.hpp"
+#include "src/restruct_poc/math/VectorOperations.hpp"
 #include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
