+++ src/restruct_poc/system/CalculateStrain.hpp
@@ -4,7 +4,7 @@
 #include <KokkosBlas.hpp>
 #include <Kokkos_Core.hpp>
 
-#include "src/restruct_poc/QuaternionOperations.hpp"
+#include "src/restruct_poc/math/QuaternionOperations.hpp"
 #include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
